requirements.txt
libnl >= 0.2.0
flask_fat @ git+https://github.com/ProjectVellum/flask-api-template.git@v0.1.2
python3-alpaka @ git+https://github.hpe.com/atsugami-kun/python3-alpaka.git@zach_dev